+++ src/index.js
@@ -1,19 +1,11 @@
-import app from './app.js';
-import express from 'express';
-import route from './routes/index';
-import 'dotenv/config';
+import app from './app';
+import config from './config/config';
 
-// const app = express();
+const currentConfig = config[process.env.NODE_ENV];
+const {port} = currentConfig;
 
-app.use(express.urlencoded({extended: true}));
-app.use(express.json());
+const server = app.listen(port, () =>
+  console.log(`App listening on ${port}!....`),
+);
 
-route(app);
-
-// const PORT =3000;
-const PORT = process.env.PORT;
-app.listen(PORT, () => {
-  console.log('Server listening on port ' + PORT);
-});
-
-
+export default server;
