--- src/middleware/hash-password.js
@@ -1,9 +1,0 @@
-import argon from 'argon2';
-
-export const hash = async (password) => {
-  return await argon.hash(password);
-};
-
-export const verify = async (hashedPassword, plainPassword) => {
-  return await argon.verify(hashedPassword, plainPassword);
-};
