+++ src/tests/auth.test.js
@@ -29,11 +29,9 @@ describe('TESTING USER AUTHENTICATION', () => {
           password: '&HrvuXn1EMe@6h1',
         })
         .end((error, response) => {
-          chai.expect(response.statusCode).to.equal(401);
-          chai.expect(response.body).to.have.property('error');
           chai
               .expect(response.body.message)
-              .to.equal('Invalid email');
+              .to.equal('Incorrect email or password');
         });
     done();
   });
@@ -42,33 +40,13 @@ describe('TESTING USER AUTHENTICATION', () => {
         .request(server)
         .post('/api/v1/users/login')
         .send({
-          email: 'arthur12kk@gmail.com',
-          password: '&HrvuXn1EMe@6h1ff',
+          email: 'arthur12@gmail.com',
+          password: '&HrvuXn1EMe@6h1u',
         })
-        .end((error, response) => {
-          chai.expect(response.statusCode).to.equal(401);
-          chai.expect(response.body).to.have.property('error');
+        .end((response) => {
           chai
               .expect(response.body.message)
-              .to.equal('Invalid password');
-        });
-    done();
-  });
-
-  it('User login fail(message:Please provide email and password!)', (done) => {
-    chai
-        .request(server)
-        .post('/api/v1/users/login')
-        .send({
-          email: '',
-          password: '',
-        })
-        .end((error, response) => {
-          chai.expect(response.status).to.equal(400);
-          chai.expect(response.body).to.have.property('error');
-          chai
-              .expect(response.body.message)
-              .to.equal('Please provide email and password!');
+              .to.equal('Incorrect email or password');
         });
     done();
   });
